+++ src/main/java/tn/esprit/spring/entity/Claims.java
@@ -1,9 +1,7 @@
 package tn.esprit.spring.entity;
 
 import java.io.Serializable;
-
-import java.util.Date;
-
+import java.sql.Date;
 
 import javax.persistence.Entity;
 import javax.persistence.EnumType;
@@ -11,84 +9,35 @@ import javax.persistence.Enumerated;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-
-
-import javax.persistence.ManyToOne;
-
-
-
-
 import javax.persistence.Temporal;
 import javax.persistence.TemporalType;
-
-
-
-
 
 @SuppressWarnings("serial")
 @Entity
 
 
 public class Claims implements Serializable {
-	
+
 	@Id
 	@GeneratedValue (strategy = GenerationType.IDENTITY)
 	
-	private Long idClaims;
+	private Double Claim_id;
 	@Temporal(TemporalType.DATE)
 	private Date Claim_date;
-	private String Claim_description; 
-	private Integer Claim_process;
 	@Enumerated(EnumType.STRING)
-	private Claim_type claim_type;
-	
-
-
-	public Long getIdClaims() {
-		return idClaims;
+	private State_claim state_claim;
+	public Double getClaim_id() {
+		return Claim_id;
 	}
-
-	public void setIdClaims(Long idClaims) {
-		this.idClaims = idClaims;
+	public void setClaim_id(Double claim_id) {
+		Claim_id = claim_id;
 	}
-
 	public Date getClaim_date() {
 		return Claim_date;
 	}
-
 	public void setClaim_date(Date claim_date) {
 		Claim_date = claim_date;
 	}
-
-	public String getClaim_description() {
-		return Claim_description;
-	}
-
-	public void setClaim_description(String claim_description) {
-		Claim_description = claim_description;
-	}
-
-	public Integer getClaim_process() {
-		return Claim_process;
-	}
-
-
-
-
-
-
-	public void setClaim_process(Integer claim_process) {
-		Claim_process = claim_process;
-	}
-
-
-
-
-
-
-	@ManyToOne
-	private User user; 
 	
-
-
 }
+
