+++ src/main/java/tn/esprit/spring/entity/State_claim.java
@@ -1,0 +1,6 @@
+package tn.esprit.spring.entity;
+
+public enum State_claim {
+	Accepted,Resiliated,Pending_Resiliation
+
+}
